+++ lib/models/app_user_preferences.dart
@@ -1,5 +1,6 @@
 import 'package:supa_architecture/json/json.dart';
 import 'package:supa_architecture/models/enum_model.dart';
+import 'package:supa_architecture/models/models.dart';
 
 mixin AppUserPreferences {
   List<JsonField> get preferenceFields => [
@@ -16,19 +17,19 @@ mixin AppUserPreferences {
       ];
 
   /// The user's admin type as an [EnumModel].
-  JsonObject<EnumModel> adminType = JsonObject<EnumModel>("adminType");
+  JsonObject<AdminType> adminType = JsonObject<AdminType>("adminType");
 
   /// The ID of the user's admin type.
   JsonInteger adminTypeId = JsonInteger("adminTypeId");
 
   /// The user's gender as an [EnumModel].
-  JsonObject<EnumModel> gender = JsonObject<EnumModel>("gender");
+  JsonObject<Gender> gender = JsonObject<Gender>("gender");
 
   /// The ID of the user's gender.
   JsonInteger genderId = JsonInteger("genderId");
 
   /// The user's language as an [EnumModel].
-  JsonObject<EnumModel> language = JsonObject<EnumModel>("language");
+  JsonObject<Language> language = JsonObject<Language>("language");
 
   /// The ID of the user's language.
   JsonInteger languageId = JsonInteger("languageId");
@@ -41,7 +42,7 @@ mixin AppUserPreferences {
       JsonBoolean("receivingSystemNotification");
 
   /// The user's timezone as an [EnumModel].
-  JsonObject<EnumModel> timezone = JsonObject<EnumModel>("timezone");
+  JsonObject<Timezone> timezone = JsonObject<Timezone>("timezone");
 
   /// The ID of the user's timezone.
   JsonInteger timezoneId = JsonInteger("timezoneId");