+++ lib/models/language_filter.dart
@@ -1,0 +1,3 @@
+import 'package:supa_architecture/models/enum_model_filter.dart';
+
+class LanguageFilter extends EnumModelFilter {}